# Speed up expand_directories function