--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,6 +1,6 @@
 [tool.poetry]
 name = "misspelling"
-version = "2.0.5"
+version = "2.0.6"
 description = "This is a Python library and tool to check for misspelled words in source code."
 authors = ["Patrick Mazulo <pmazulo@gmail.com>"]
 license = "GNU GENERAL PUBLIC LICENSE"