--- a/src/misspelling_lib/utils/files.py
+++ b/src/misspelling_lib/utils/files.py
@@ -7,7 +7,7 @@ from typing import Iterator, List, NoReturn, Union
 EXCLUDED_FILES_RE = re.compile(r"\.(pyc|s?o|a|sh|txt|coverage|gitignore|python-version)|LICENSE$")
 EXCLUDED_DIRS_RE = re.compile(
     r"\.(git|github|mypy_cache|pytest_cache|idea|vscode)|"
-    r"^(\*egg|\*egg-info|CVS|bin|node_modules|json_sources|tests/sources)$"
+    r"(\*egg|\*egg-info|CVS|bin|node_modules|__pycache__|json_sources|test_assets)"
 )
 
 
@@ -28,10 +28,9 @@ def parse_file_list(filename: Path) -> Union[List[Path], NoReturn]:
 def expand_directories(path_list: List[Path]) -> Iterator[Path]:
     """Return list with directories replaced their contained files."""
     for path in path_list:
-        if path.is_dir() and not EXCLUDED_DIRS_RE.match(path.as_posix()):
-            for filenames in iglob(path.as_posix()):
-                for name in filenames:
-                    if not EXCLUDED_FILES_RE.search(name):
-                        yield Path(name)
+        if path.is_dir() and not EXCLUDED_DIRS_RE.search(path.as_posix()):
+            for filename in iglob(f"{path}/**/*.*", recursive=True):
+                if not EXCLUDED_DIRS_RE.search(filename) and not EXCLUDED_FILES_RE.search(filename):
+                    yield Path(filename)
         else:
             yield path