--- a/src/misspelling_lib/utils/version.py
+++ b/src/misspelling_lib/utils/version.py
@@ -1,2 +1,2 @@
 def get_version() -> str:
-    return "2.0.5"
+    return "2.0.6"