--- a/tests/assets/good_file_list
+++ b/tests/assets/good_file_list
@@ -1,1 +1,0 @@
-assets/nine_misspellings.c