--- a/tests/test_assets/good_file_list
+++ b/tests/test_assets/good_file_list
@@ -1,0 +1,1 @@
+test_assets/nine_misspellings.c