In a directory with 1411 valid files, the new implementation using [`os.scandir`](https://docs.python.org/3/library/os.html#os.scandir) took simply 8ms. The previous implementation for the same project would found more files (6692) but most of them not valid ones.

Here are the results using different implementations:

```bash
❯ python second_performance_test.py
os.walk                 took 20 ms. Found files: 1411
os.walk-glob    took 30 ms. Found files: 1399
glob.glob               took 26 ms. Found files: 1385
glob.iglob              took 27 ms. Found files: 1385
pathlib.rglob   took 47 ms. Found files: 1406
find_files              took 39 ms. Found files: 1411
fast_scandir    took 8 ms. Found files: 1411
run_previous_os_walk_from_the_project   took 22 ms. Found files: 6692
```

and here is the code for this: https://gist.github.com/mazulo/f1fe5047fda3ab3a8f965ddfb7e39f2a